Seperated user,customer,employee classes
